Normalized redux aliases and namespaces
