+++ package.json
@@ -31,7 +31,6 @@
     "formik": "^2.2.9",
     "hypercore": "^10.0.0-alpha.18",
     "lodash": "^4.17.21",
-    "moment": "^2.29.4",
     "nodejs-mobile-react-native": "^0.6.2",
     "patch-package": "^6.4.7",
     "postinstall-postinstall": "^2.1.0",
