+++ src/store/contacts.ts
@@ -1,5 +1,5 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { accountLogout } from './account';
+import { createSlice } from '@reduxjs/toolkit';
+import { accountLogout } from './thunks/accountLogout';
 
 interface ContactsState {}
 
@@ -12,8 +12,7 @@ export const contactsSlice = createSlice({
   extraReducers: builder => {
     // clear state on logout
     builder.addCase(accountLogout.fulfilled, () => {
-      const newState = { ...initialState };
-      return newState;
+      return { ...initialState };
     });
   },
 });
